refactor(selectable-search-input): fix eslint warnings
